Rename sys_render2_* → sys_render_*
