+++ DeferredShading/game.ts
@@ -16,9 +16,9 @@ import {sys_camera} from "./systems/sys_camera.js";
 import {sys_control_always} from "./systems/sys_control_always.js";
 import {sys_light} from "./systems/sys_light.js";
 import {sys_move} from "./systems/sys_move.js";
-import {sys_render_deferred} from "./systems/sys_render2_deferred.js";
-import {sys_render_depth} from "./systems/sys_render2_depth.js";
-import {sys_render_postprocess} from "./systems/sys_render2_postprocess.js";
+import {sys_render_deferred} from "./systems/sys_render_deferred.js";
+import {sys_render_depth} from "./systems/sys_render_depth.js";
+import {sys_render_postprocess} from "./systems/sys_render_postprocess.js";
 import {sys_resize} from "./systems/sys_resize.js";
 import {sys_transform} from "./systems/sys_transform.js";
 import {World} from "./world.js";
