--- DeferredShading/systems/sys_render2_depth.ts
@@ -1,1 +1,0 @@
-../../core/systems/sys_render2_depth.ts
