--- DeferredShading/systems/sys_render2_postprocess.ts
@@ -1,1 +1,0 @@
-../../core/systems/sys_render2_postprocess.ts
