+++ DeferredShading/systems/sys_render_postprocess.ts
@@ -1,0 +1,1 @@
+../../core/systems/sys_render_postprocess.ts
