+++ FlatShading/world.ts
@@ -4,7 +4,7 @@ import {Children} from "./components/com_children.js";
 import {ControlAlways} from "./components/com_control_always.js";
 import {Light} from "./components/com_light.js";
 import {Move} from "./components/com_move.js";
-import {Render} from "./components/com_render2.js";
+import {Render} from "./components/com_render.js";
 import {Transform} from "./components/com_transform.js";
 
 const enum Component {
