--- ForwardShading/systems/sys_render2_forward.ts
@@ -1,1 +1,0 @@
-../../core/systems/sys_render2_forward.ts
