+++ ForwardShading/systems/sys_render_forward.ts
@@ -1,0 +1,1 @@
+../../core/systems/sys_render_forward.ts
