+++ WebXR/world.ts
@@ -3,7 +3,7 @@ import {Camera} from "./components/com_camera.js";
 import {Children} from "./components/com_children.js";
 import {ControlXr} from "./components/com_control_xr.js";
 import {Light} from "./components/com_light.js";
-import {Render} from "./components/com_render2.js";
+import {Render} from "./components/com_render.js";
 import {Transform} from "./components/com_transform.js";
 
 const enum Component {
