+++ core/components/com_render.ts
